+++ lib/constants/changelog.dart
@@ -5,8 +5,12 @@ const changelog = [
     'titles': [],
     'added': [
       'Select a custom range in the Summary',
+      'Predefined income or expense categories',
+      'About section in the Settings page',
     ],
-    'fixed': [],
+    'fixed': [
+      'Category or Account can\'t be deleted if they contain any transaction'
+    ],
   },
   {
     'version': 'v4.4.0',